requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.6
mangum
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cachetools==5.3.3
pydantic-settings==2.2.1
